Current progress

+++ Coursera/Yandex_CppWhite/Coursera_CppWhite_Week01_Task06.cpp
@@ -15,7 +15,7 @@ int find2nd_f()
    {
       if (s[i] == 'f')
       {
-         index = i;
+         index = static_cast<int>(i);
          result++;
       }
       if (result == 0)
