+++ Coursera/Yandex_CppWhite/Coursera_CppWhite_Week02_Task01.cpp
@@ -4,8 +4,8 @@ using namespace std;
 
 bool IsPalindrom21(const string& s)
 {
-	int b = 0;
-	int e = s.size() - 1;
+	size_t b = 0;
+	size_t e = s.size() - 1;
 
 	while (b < e)
 	{
