+++ Coursera/Yandex_CppWhite/Coursera_CppWhite_Week02_Task06.cpp
@@ -4,8 +4,8 @@ using namespace std;
 
 void Reverse(vector<int>& v)
 {
-   int b = 0;
-   int e = v.size() - 1;
+   size_t b = 0;
+   size_t e = v.size() - 1;
    while (b < e)
    {
       swap(v[b++], v[e--]);
