+++ Coursera/Yandex_CppWhite/Coursera_CppWhite_Week02_Task07.cpp
@@ -4,9 +4,9 @@ using namespace std;
 
 vector<int> Reversed(const vector<int>& v)
 {
-   int sz = v.size();
+   size_t sz = v.size();
    vector<int> r(sz);
-   for(int i = sz - 1; i >=0; i--)
+   for(size_t i = sz - 1; i >=0; i--)
    {
       r[sz - i - 1] = v[i];
    }
